Add NeCTAR gitreview

+++ openstack_dashboard/dashboards/admin/images/views.py
@@ -101,7 +101,7 @@ class IndexView(tables.DataTableView):
                 invalid_msg = ('API query is not valid and is ignored: %s=%s'
                                % (filter_field, filter_string))
                 try:
-                    filter_string = long(float(filter_string) * (units.Mi))
+                    filter_string = int(float(filter_string) * (units.Mi))
                     if filter_string >= 0:
                         filters[filter_field] = filter_string
                     else:
